Revert back to bundled ICU
Maintaining compatibility with ICU 63 is causing issues with certain
pages, such as that reported in the fixed issue below. The culprit seems
to be patch buster/icu63.patch, but I have not diagnosed the issue
enough to fix the patch, so I'm reverting back to the bundled ICU for
now.

Fixes https://github.com/Eloston/ungoogled-chromium/issues/778