Optimize removal of files returned by mk-origtargz
The original version was performing a fork and exec combination
for every file which gets very expensive as the number of files
increases. This new version offloads the task to xargs which can
make a more informed decision about how best to dispatch the task
to the rm command. This should result in a reduction in how long
the get-orig-source target takes to run, resulting in around 10%
faster builds of this target.